XFA: merge patch from CL 792953005, fix most warnings
Includes fixes to XFA specific warnings -- benign truncations.
Bug https://code.google.com/p/pdfium/issues/detail?id=104
was filed to track changing types to avoid some truncations.

Resolve all but two VC++ build warnings in pdfium.

pdfium builds on Win32 have about 85 warnings (250 in the XFA
branch, totaling over 480 lines!), mostly from four lines in
a header file and a warning that should be disabled. This
change resolves all but two of them and turns on
warning-as-errors. Bugs have been filed for the two
remaining warnings:
https://code.google.com/p/pdfium/issues/detail?id=100
the 64-bit warnings:
https://code.google.com/p/pdfium/issues/detail?id=101
and the Linux warnings:
https://code.google.com/p/pdfium/issues/detail?id=102

The fix to the double->float truncation bugs will also
improve code-generation.

R=bo_xu@foxitsoftware.com, tsepez@chromium.org

Review URL: https://codereview.chromium.org/792953005

BUG= https://code.google.com/p/pdfium/issues/detail?id=100

Review URL: https://codereview.chromium.org/834413002

+++ fpdfsdk/include/fsdk_mgr.h
@@ -476,10 +476,10 @@ public:
 			double bottom;
 			m_pInfo->FFI_GetPageViewRect(m_pInfo, page, &left, &top, &right, &bottom);
 
-			dstRect.left = left;
-			dstRect.top = top < bottom? bottom:top;
-			dstRect.bottom = top < bottom? top:bottom;
-			dstRect.right = right;
+			dstRect.left = static_cast<float>(left);
+			dstRect.top = static_cast<float>(top < bottom ? bottom : top);
+			dstRect.bottom = static_cast<float>(top < bottom ? top : bottom);
+			dstRect.right = static_cast<float>(right);
 		}
 	}
